Move checkFormSubmissionPeriodAndBapStatus() out of formio2022.js file and into bap.js utility file
